functions/requirements.txt
pinecone-client
pandas
tenacity
transformers
--extra-index-url https://download.pytorch.org/whl/cpu
torch